+++ libs/pika/async_mpi/CMakeLists.txt
@@ -13,12 +13,12 @@ endif()
 
 # Default location is $PIKA_ROOT/libs/mpi/include
 set(async_mpi_headers
-    pika/async_mpi/mpi_exception.hpp pika/async_mpi/mpi_future.hpp
+    pika/async_mpi/mpi_exception.hpp pika/async_mpi/mpi_polling.hpp
     pika/async_mpi/transform_mpi.hpp
 )
 
 # Default location is $PIKA_ROOT/libs/mpi/src
-set(mpi_sources mpi_exception.cpp mpi_future.cpp)
+set(mpi_sources mpi_exception.cpp mpi_polling.cpp)
 
 include(pika_add_module)
 pika_add_module(
